Small modification to weatherInformation class

+++ STEVE_Chase_Car/STEVE_Chase_Car/Code/weatherInformation.cs
@@ -23,7 +23,7 @@ namespace STEVE_Chase_Car.Code
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
+        public double pressure { get; set; }
         public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
@@ -32,7 +32,7 @@ namespace STEVE_Chase_Car.Code
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
